requirements-scraper.txt
# Core scraping dependencies
scrapy>=2.11.0
itemloaders>=1.1.0
w3lib>=2.1.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Firebase integration
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0

# Google Sheets integration
gspread>=5.10.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# HTTP requests and parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# Date handling
python-dateutil>=2.8.0

# Async/await support
twisted>=22.10.0

# Environment variables
python-dotenv>=1.0.0

# Proxy rotation (optional)
scrapy-rotating-proxies>=0.6.2

# User agent rotation
scrapy-user-agents>=0.1.1

# Selenium for JavaScript rendering (optional)
selenium>=4.11.0
webdriver-manager>=3.9.0

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.7.0
flake8>=6.0.0
